test: update tests to work with cypress
