+++ test/integration/unit/functions/explorer.test.ts
@@ -1,0 +1,25 @@
+import { ChainId } from '@sushiswap/sdk'
+import { getExplorerLink } from '../../../../src/functions/explorer'
+
+describe('utils', () => {
+  describe('#getExplorerLink', () => {
+    it('correct for tx', () => {
+      expect(getExplorerLink(1, 'abc', 'transaction')).to.equal('https://etherscan.io/tx/abc')
+    })
+    it('correct for token', () => {
+      expect(getExplorerLink(1, 'abc', 'token')).to.equal('https://etherscan.io/token/abc')
+    })
+    it('correct for address', () => {
+      expect(getExplorerLink(1, 'abc', 'address')).to.equal('https://etherscan.io/address/abc')
+    })
+    /*it('unrecognized chain id defaults to mainnet', () => {
+      expect(getExplorerLink(2, 'abc', 'address')).to.equal('https://etherscan.io/address/abc')
+    })*/
+    it('ropsten', () => {
+      expect(getExplorerLink(3, 'abc', 'address')).to.equal('https://ropsten.etherscan.io/address/abc')
+    })
+    it('enum', () => {
+      expect(getExplorerLink(ChainId.RINKEBY, 'abc', 'address')).to.equal('https://rinkeby.etherscan.io/address/abc')
+    })
+  })
+})
