+++ test/integration/unit/functions/format.test.ts
@@ -1,0 +1,19 @@
+import { shortenAddress } from '../../../../src/functions/format'
+
+describe('#shortenAddress', () => {
+  it('throws on invalid address', () => {
+    expect(() => shortenAddress('abc')).to.throw("Invalid 'address'")
+  })
+
+  it('truncates middle characters', () => {
+    expect(shortenAddress('0xf164fc0ec4e93095b804a4795bbe1e041497b92a')).to.equal('0xf164...b92a')
+  })
+
+  it('truncates middle characters even without prefix', () => {
+    expect(shortenAddress('f164fc0ec4e93095b804a4795bbe1e041497b92a')).to.equal('0xf164...b92a')
+  })
+
+  it('renders checksummed address', () => {
+    expect(shortenAddress('0x2E1b342132A67Ea578e4E3B814bae2107dc254CC'.toLowerCase())).to.equal('0x2E1b...54CC')
+  })
+})
